+++ packages/api/lib/testUtils.js
@@ -19,12 +19,10 @@ const {
   translateApiExecutionToPostgresExecution,
   translateApiProviderToPostgresProvider,
   translateApiRuleToPostgresRuleRaw,
-  translatePostgresPdrToApiPdr,
   translatePostgresRuleToApiRule,
 } = require('@cumulus/db');
 const {
   indexRule,
-  indexPdr,
   deleteExecution,
 } = require('@cumulus/es-client/indexer');
 const {
@@ -562,8 +560,6 @@ const createPdrTestRecords = async (context, pdrParams = {}) => {
   const {
     knex,
     pdrPgModel,
-    esClient,
-    esPdrsClient,
     collectionCumulusId,
     providerCumulusId,
   } = context;
@@ -585,14 +581,8 @@ const createPdrTestRecords = async (context, pdrParams = {}) => {
   const originalPgRecord = await pdrPgModel.get(
     knex, { cumulus_id: pgPdr.cumulus_id }
   );
-  const originalPdr = await translatePostgresPdrToApiPdr(originalPgRecord, knex);
-  await indexPdr(esClient, originalPdr, process.env.ES_INDEX);
-  const originalEsRecord = await esPdrsClient.get(
-    originalPdr.pdrName
-  );
   return {
     originalPgRecord,
-    originalEsRecord,
   };
 };